requirements.txt
fastapi[all]
pydantic
uvicorn
pygithub
codegrade